Move exceptions to the `org.spine3.base` package.
