+++ client/src/test/java/org/spine3/base/StringifiersShould.java
@@ -24,7 +24,6 @@ import com.google.common.testing.NullPointerTester;
 import com.google.protobuf.Timestamp;
 import com.google.protobuf.util.Timestamps;
 import org.junit.Test;
-import org.spine3.base.error.IllegalConversionArgumentException;
 import org.spine3.test.identifiers.IdWithPrimitiveFields;
 import org.spine3.test.types.Task;
 import org.spine3.type.ParametrizedTypeImpl;
